Create analog folder for analog routines
- ADC and DAC tested and working.
- DAC is showing voltage in mV. Still need to add a NTC handler
