+++ src/main.c
@@ -14,6 +14,7 @@
 #include "global.h"
 #include "menu.h"
 #include "ctype.h"
+#include "analog.h"
 
 #define COUNTER_1_US 13
 
@@ -29,20 +30,11 @@
 #define N_ROWS 4
 #define N_COLS 4
 
-//Analog voltage reference in milivolts
-#define ANALOG_VREF_MV 3300
-
-//DAC resolutions. Value of 2^n
-#define DAC_RES_BITS 4096
-
-//ADC resolutions. Value of 2^n
-#define ADC_RES_BITS 4096
-
 //Number of samples the ADC must take before calculating the analog value.
-#define ADC_SAMPLES 8
+#define ADC_SAMPLES 16
 
 //Time in ms between ADC samples.
-#define ADC_SAMPLES_INTERVAL 10
+#define ADC_SAMPLES_INTERVAL 5
 
 /****************************
  * Global vars and structs
@@ -102,7 +94,6 @@ char stringToPrint[16] = "                ";
 /*******************************
  * Function prototypes
  *******************************/
-void writeDAC(uint16_t valueInMv);
 void initGPIO();
 void backlightOn();
 void backlightOff();
@@ -114,8 +105,6 @@ void counterTask();
 void analogTask();
 uint32_t getGlobalSystickValue();
 void taskHandler();
-void writeDAC(uint16_t valueInMv);
-void readADCCh7(uint16_t *value);
 
 /*******************************
  * End function prototypes
@@ -126,6 +115,8 @@ int main(void)
 	SysTick_Config(SystemCoreClock / 1000);
 
 	initGPIO();
+	initDAC();
+	initADC();
 
 	LCDTaskData.intervalInMs = 250;
 	keypadTaskData.intervalInMs = 10;
@@ -150,8 +141,6 @@ void initGPIO()
 	RCC_AHB1PeriphClockCmd(RCC_AHB1Periph_GPIOC, ENABLE);
 
 	GPIO_InitTypeDef GPIO_InitStruct;
-	DAC_InitTypeDef DAC_InitStruct;
-	ADC_InitTypeDef ADC_InitStructure;
 
 	//Inicializo BACKLIGHT
 	GPIO_InitStruct.GPIO_Pin = BACKLIGHT_PIN;
@@ -170,44 +159,6 @@ void initGPIO()
 	keypadInitStruct.keyMapping = keypadKeyMapping;
 
 	keypadInit(keypadInitStruct);
-
-	//Init DAC GPIO A4
-	GPIO_InitStruct.GPIO_Pin = GPIO_Pin_4;
-	GPIO_InitStruct.GPIO_Mode = GPIO_Mode_AN;
-	GPIO_InitStruct.GPIO_PuPd = GPIO_PuPd_NOPULL;
-	GPIO_Init(GPIOA, &GPIO_InitStruct);
-
-	RCC_APB1PeriphClockCmd(RCC_APB1Periph_DAC, ENABLE);
-
-	DAC_InitStruct.DAC_Trigger = DAC_Trigger_None;
-	DAC_InitStruct.DAC_WaveGeneration = DAC_WaveGeneration_None;
-	DAC_InitStruct.DAC_OutputBuffer = DAC_OutputBuffer_Enable;
-	DAC_Init(DAC_Channel_1, &DAC_InitStruct);
-
-	DAC_Cmd(DAC_Channel_1, ENABLE);
-
-	//Init ADC 1 / GPIOA 7 Channel 7
-	GPIO_InitStruct.GPIO_Pin = GPIO_Pin_7;
-	GPIO_InitStruct.GPIO_Mode = GPIO_Mode_AN;
-	GPIO_InitStruct.GPIO_PuPd = GPIO_PuPd_NOPULL;
-	GPIO_Init(GPIOA, &GPIO_InitStruct);
-
-	RCC_APB2PeriphClockCmd(RCC_APB2Periph_ADC1, ENABLE);
-
-	ADC_StructInit(&ADC_InitStructure);
-	ADC_InitStructure.ADC_Resolution = ADC_Resolution_12b;
-	ADC_InitStructure.ADC_ScanConvMode = ENABLE;
-	ADC_InitStructure.ADC_ContinuousConvMode = DISABLE;
-	ADC_InitStructure.ADC_ExternalTrigConvEdge = ADC_ExternalTrigConvEdge_None;
-	ADC_InitStructure.ADC_DataAlign = ADC_DataAlign_Right;
-	ADC_InitStructure.ADC_NbrOfConversion = 1;
-	ADC_Init(ADC1, &ADC_InitStructure);
-
-	ADC_RegularChannelConfig(ADC1, ADC_Channel_7, 1, ADC_SampleTime_3Cycles);
-
-	//Enable ADC
-	ADC_Cmd(ADC1, ENABLE);
-
 
 	LCD_init();
 	backlightOn();
@@ -324,10 +275,14 @@ void keypadTask()
 				default:
 					if (isdigit(pressedKey))
 					{
-						//I should ask for the current menu level before setting up the DAC value.
-						DACSelectedValueInMv = (pressedKey - 48) * 100;
-						if (DACSelectedValueInMv != DACCurrentValueInMv)
-							menuSetDACValue(DACSelectedValueInMv);
+						//This must be improved in order to not use literals
+						if (getMenuCurrentLevel() == 1 && getMenuCurrentItem() == 3)
+						{
+							DACSelectedValueInMv = (pressedKey - 48) * 100;
+							if (DACSelectedValueInMv != DACCurrentValueInMv)
+								menuSetDACValue(DACSelectedValueInMv);
+						}
+
 					}
 				break;
 			}
@@ -432,31 +387,6 @@ uint32_t getGlobalSystickValue()
 }
 /****************************
  * End timing functions
- ****************************/
-
-/****************************
- * Analog functions
- ****************************/
-
-void writeDAC(uint16_t valueInMv)
-{
-	//This is the minimum change in milivolts handled by the DAC.
-	float DACResolutionInMv = (float)ANALOG_VREF_MV / (float)DAC_RES_BITS;
-
-	uint16_t DACValue = valueInMv / DACResolutionInMv;
-	DAC_SetChannel1Data(DAC_Align_12b_R, DACValue);
-}
-
-void readADCCh7(uint16_t *value)
-{
-	ADC_SoftwareStartConv(ADC1);
-
-	while (ADC_GetFlagStatus(ADC1, ADC_FLAG_EOC) == RESET);
-	*value = ADC_GetConversionValue(ADC1);
-}
-			
-/****************************
- * End analog functions
  ****************************/
 
 /****************************
